Move domain packages to another package called domains
