--- src/main/java/br/com/elo7/sonda/candidato/SevenMarsApplication.java
@@ -1,11 +1,0 @@
-package br.com.elo7.sonda.candidato;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SevenMarsApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(SevenMarsApplication.class, args);
-	}
-}
