+++ src/main/java/br/com/elo7/sonda/candidato/StarlinkApplication.java
@@ -1,0 +1,11 @@
+package br.com.elo7.sonda.candidato;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StarlinkApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(StarlinkApplication.class, args);
+	}
+}
