+++ src/main/java/br/com/elo7/sonda/candidato/domains/planet/service/PlanetService.java
@@ -1,0 +1,11 @@
+package br.com.elo7.sonda.candidato.domains.planet.service;
+
+import br.com.elo7.sonda.candidato.domains.planet.dto.PlanetDTO;
+
+import java.util.List;
+
+public interface PlanetService {
+    List<PlanetDTO> findAll();
+    PlanetDTO find(Long id);
+    void save(PlanetDTO planetDTO);
+}
