+++ src/main/java/br/com/elo7/sonda/candidato/domains/probe/dto/AreaDTO.java
@@ -1,0 +1,31 @@
+package br.com.elo7.sonda.candidato.domains.probe.dto;
+
+public class AreaDTO {
+    private int x;
+    private int y;
+    private Long planetId;
+
+    public int getX() {
+        return x;
+    }
+
+    public void setX(int x) {
+        this.x = x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setY(int y) {
+        this.y = y;
+    }
+
+    public Long getPlanetId() {
+        return planetId;
+    }
+
+    public void setPlanetId(Long planetId) {
+        this.planetId = planetId;
+    }
+}
