+++ src/main/java/br/com/elo7/sonda/candidato/domains/probe/dto/ProbeDTO.java
@@ -1,0 +1,60 @@
+package br.com.elo7.sonda.candidato.domains.probe.dto;
+
+import br.com.elo7.sonda.candidato.domains.probe.enumeration.Direction;
+
+public class ProbeDTO {
+    private Long id;
+    private String name;
+    private String description;
+    private Direction direction;
+    private int x;
+    private int y;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public Direction getDirection() {
+        return direction;
+    }
+
+    public void setDirection(Direction direction) {
+        this.direction = direction;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public void setX(int x) {
+        this.x = x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setY(int y) {
+        this.y = y;
+    }
+}
