+++ src/main/java/br/com/elo7/sonda/candidato/domains/probe/entity/Probe.java
@@ -1,0 +1,79 @@
+package br.com.elo7.sonda.candidato.domains.probe.entity;
+
+import br.com.elo7.sonda.candidato.domains.probe.enumeration.Direction;
+import com.sun.istack.NotNull;
+
+import javax.persistence.*;
+
+@Entity
+public class Probe {
+
+    @Id
+    @Column
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column
+    @NotNull
+    private String name;
+
+    @Column
+    private String description;
+
+    @Column
+    private int x;
+
+    @Column
+    private int y;
+
+    @Column
+    private Direction direction;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public void setX(int x) {
+        this.x = x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setY(int y) {
+        this.y = y;
+    }
+
+    public Direction getDirection() {
+        return direction;
+    }
+
+    public void setDirection(Direction direction) {
+        this.direction = direction;
+    }
+}
