+++ src/main/java/br/com/elo7/sonda/candidato/domains/probe/enumeration/Direction.java
@@ -1,0 +1,9 @@
+package br.com.elo7.sonda.candidato.domains.probe.enumeration;
+
+public enum Direction {
+	N('N'), W('W'), E('E'), S('S');
+	public final char direction;
+	Direction(char direction) {
+		this.direction = direction;
+	}
+}
