--- src/main/java/br/com/elo7/sonda/candidato/planet/entity/Object.java
@@ -1,68 +1,0 @@
-package br.com.elo7.sonda.candidato.planet.entity;
-
-import javax.persistence.*;
-
-@Entity
-public class Object {
-    @Id
-    @Column
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String description;
-    @Column
-    private int x;
-    @Column
-    private int y;
-    @ManyToOne
-    @JoinColumn(name = "planet_id")
-    private Planet planet;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public int getX() {
-        return x;
-    }
-
-    public void setX(int x) {
-        this.x = x;
-    }
-
-    public int getY() {
-        return y;
-    }
-
-    public void setY(int y) {
-        this.y = y;
-    }
-
-    public Planet getPlanet() {
-        return planet;
-    }
-
-    public void setPlanet(Planet planet) {
-        this.planet = planet;
-    }
-}
