--- src/main/java/br/com/elo7/sonda/candidato/planet/service/ObjectService.java
@@ -1,9 +1,0 @@
-package br.com.elo7.sonda.candidato.planet.service;
-
-import br.com.elo7.sonda.candidato.planet.dto.ObjectDTO;
-import br.com.elo7.sonda.candidato.planet.entity.Object;
-
-public interface ObjectService {
-    void receiveObject(ObjectDTO objectDTO);
-    void updateCoordinates(Object object, int x, int y);
-}
