--- src/main/java/br/com/elo7/sonda/candidato/probe/service/ProbeService.java
@@ -1,16 +1,0 @@
-package br.com.elo7.sonda.candidato.probe.service;
-
-import br.com.elo7.sonda.candidato.probe.dto.AreaDTO;
-import br.com.elo7.sonda.candidato.probe.dto.CommandDTO;
-import br.com.elo7.sonda.candidato.probe.dto.ProbeDTO;
-
-import java.util.List;
-
-public interface ProbeService {
-    List<ProbeDTO> findAll();
-    ProbeDTO find(Long id);
-    void save(ProbeDTO planetDTO);
-    void sendCommand(Long probeId, CommandDTO commandDTO);
-    void moveProbe(ProbeDTO probe, String commands);
-    void sendToPlanet(Long probeId, AreaDTO area);
-}
