--- src/main/java/br/com/elo7/sonda/candidato/probe/service/movements/Right.java
@@ -1,22 +1,0 @@
-package br.com.elo7.sonda.candidato.probe.service.movements;
-
-import br.com.elo7.sonda.candidato.probe.enumeration.Command;
-import br.com.elo7.sonda.candidato.probe.enumeration.Direction;
-import org.springframework.stereotype.Component;
-
-@Component
-public class Right extends Movement {
-    public Command findCommand() {
-        return Command.R;
-    }
-    public Movement moveTo() {
-        Direction newDirection = switch (probe.getDirection()) {
-            case N -> Direction.E;
-            case E -> Direction.S;
-            case S -> Direction.W;
-            case W -> Direction.N;
-        };
-        probe.setDirection(newDirection);
-        return this;
-    }
-}
