+++ src/test/java/br/com/elo7/sonda/candidato/planet/service/PlanetServiceTest.java
@@ -1,9 +1,11 @@
 package br.com.elo7.sonda.candidato.planet.service;
 
+import br.com.elo7.sonda.candidato.domains.planet.service.PlanetService;
+import br.com.elo7.sonda.candidato.domains.planet.service.PlanetServiceImpl;
 import br.com.elo7.sonda.candidato.exception.ApplicationException;
-import br.com.elo7.sonda.candidato.planet.dto.PlanetDTO;
-import br.com.elo7.sonda.candidato.planet.entity.Planet;
-import br.com.elo7.sonda.candidato.planet.repository.PlanetRepository;
+import br.com.elo7.sonda.candidato.domains.planet.dto.PlanetDTO;
+import br.com.elo7.sonda.candidato.domains.planet.entity.Planet;
+import br.com.elo7.sonda.candidato.domains.planet.repository.PlanetRepository;
 import org.jeasy.random.EasyRandom;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
