+++ src/test/java/br/com/elo7/sonda/candidato/probe/service/ProbeServiceTest.java
@@ -1,16 +1,18 @@
 package br.com.elo7.sonda.candidato.probe.service;
 
+import br.com.elo7.sonda.candidato.domains.probe.service.ProbeService;
+import br.com.elo7.sonda.candidato.domains.probe.service.ProbeServiceImpl;
 import br.com.elo7.sonda.candidato.exception.ApplicationException;
-import br.com.elo7.sonda.candidato.planet.service.ObjectService;
-import br.com.elo7.sonda.candidato.probe.dto.AreaDTO;
-import br.com.elo7.sonda.candidato.probe.dto.CommandDTO;
-import br.com.elo7.sonda.candidato.probe.dto.ProbeDTO;
-import br.com.elo7.sonda.candidato.probe.entity.Probe;
-import br.com.elo7.sonda.candidato.probe.repository.ProbeRepository;
-import br.com.elo7.sonda.candidato.probe.service.movements.Forward;
-import br.com.elo7.sonda.candidato.probe.service.movements.Left;
-import br.com.elo7.sonda.candidato.probe.service.movements.Movement;
-import br.com.elo7.sonda.candidato.probe.service.movements.Right;
+import br.com.elo7.sonda.candidato.domains.planet.service.ObjectService;
+import br.com.elo7.sonda.candidato.domains.probe.dto.AreaDTO;
+import br.com.elo7.sonda.candidato.domains.probe.dto.CommandDTO;
+import br.com.elo7.sonda.candidato.domains.probe.dto.ProbeDTO;
+import br.com.elo7.sonda.candidato.domains.probe.entity.Probe;
+import br.com.elo7.sonda.candidato.domains.probe.repository.ProbeRepository;
+import br.com.elo7.sonda.candidato.domains.probe.service.movements.Forward;
+import br.com.elo7.sonda.candidato.domains.probe.service.movements.Left;
+import br.com.elo7.sonda.candidato.domains.probe.service.movements.Movement;
+import br.com.elo7.sonda.candidato.domains.probe.service.movements.Right;
 import org.jeasy.random.EasyRandom;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,8 +26,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static br.com.elo7.sonda.candidato.probe.enumeration.Direction.E;
-import static br.com.elo7.sonda.candidato.probe.enumeration.Direction.N;
+import static br.com.elo7.sonda.candidato.domains.probe.enumeration.Direction.E;
+import static br.com.elo7.sonda.candidato.domains.probe.enumeration.Direction.N;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
